Added MicroCardContainer component
